pvc_scenarios porting
